+++ cement/core/backend.py
@@ -1,3 +1,3 @@
 """Cement core backend module."""
 
-VERSION = (3, 0, 8, 'final', 0)  # pragma: nocover
+VERSION = (3, 0, 10, 'final', 0)  # pragma: nocover